Fixes #668: Separate component for Skill Cards List
